requirements.txt
bardapi==0.1.33
matplotlib==3.7.2
numpy==1.25.2
pandas==2.0.3
Pillow==9.5.0
python-docx==0.8.11
seaborn==0.12.2
streamlit==1.25.0